refactor(bestax-migrate): hoist source-agnostic helpers to _shared
jsx-utils.ts and props.ts carry no react-bulma-components data — the
TransformContext bag, the AST helpers and the PropAction interpreter are
generic. Move both to src/sources/_shared/ so a second source can share
them instead of copying.

applyUniversalProps() took its table from RBC's mapping.ts by import;
it now takes it as a parameter, which is the only coupling that had to
change. Everything else moves verbatim.

responsive.ts and styles.ts stay per-source: rbx's breakpoint shape and
stylesheet story both differ from RBC's.
