+++ bestax-migrate/src/sources/react-bulma-components/specials.ts
@@ -22,7 +22,7 @@ import {
   removeAttr,
   resolveBooleanish,
   type TransformContext,
-} from './jsx-utils.js';
+} from '../_shared/jsx-utils.js';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
